requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-google-genai
langchain-community
faiss-cpu
pypdf
tqdm
